wip in changing to async interface
