+++ Cargo.toml
@@ -23,15 +23,9 @@ codecov = {repository = "Drevoed/narwhal", branch = "master", service = "github"
 serde = "1.0.98"
 serde_derive = "1.0.98"
 serde_json = "1.0.40"
-reqwest = {version = "0.9.19", optional = true}
 lazy_static = "1.3.0"
-hyper = {version = "0.12.33", optional = true}
+hyper = "0.12.33"
+hyper-tls = "0.3.2"
 log = "0.4.8"
-snafu = "0.4.4"
-
-[features]
-default = ["synchronous"]
-
-synchronous = ["reqwest"]
-
-asynchronous = ["hyper"]
+snafu = { version = "0.4.4", features = ["futures-01"]}
+futures = "0.1.28"
