+++ src/synchronous/mod.rs
@@ -1,4 +1,2 @@
 pub mod client;
 pub mod ddragon;
-
-pub use client::LeagueAPI;
