+++ src/types.rs
@@ -1,1 +1,2 @@
-
+use crate::error::ClientError;
+use futures::prelude::*;
